% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_coef.R
\name{\%r\%}
\alias{\%r\%}
\title{Compute regression coefficients}
\usage{
y \%r\% x
}
\arguments{
\item{y}{Response \code{vector}.}

\item{x}{Design \code{matrix} or vector.}
}
\description{
Compute regression coefficients
}
\details{
Uses \link[stats]{lm} then \link[stats]{coef}.
}
\examples{
## linear regression
cars$speed \%r\% cars$dist
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[stats]{coef}}
}
