% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snipes.R
\docType{data}
\name{snipes}
\alias{snipes}
\title{Snipes price data}
\format{
A data frame with 48 rows and 3 columns:
\describe{
\item{discount}{Discounted price of sneakers}
\item{brand}{Brand of sneakers}
\item{price}{Original price of sneakers}
}
}
\source{
\url{https://www.snipes.ch/}
}
\usage{
snipes
}
\description{
Snipes price data
}
\keyword{datasets}
